+++ addOns/sequence/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Initial sequence-activeScan implementation.
+- Data for reporting.
+
 ### Changed
 - Update minimum ZAP version to 2.15.0.
 - Maintenance changes.